Rename state-token-generator to random-string-generator
