+++ src/containers/App/saga.test.js
@@ -9,7 +9,7 @@ import CONFIGURATION from 'shared/services/configuration/configuration';
 import { authCall } from 'shared/services/api/api';
 import { login, logout } from 'shared/services/auth/auth';
 import fileUploadChannel from 'shared/services/file-upload-channel';
-import stateTokenGenerator from 'shared/services/auth/services/state-token-generator/state-token-generator';
+import randomStringGenerator from 'shared/services/auth/services/random-string-generator/random-string-generator';
 import { VARIANT_ERROR, TYPE_GLOBAL } from 'containers/Notification/constants';
 import userJson from 'utils/__tests__/fixtures/user.json';
 
@@ -39,7 +39,7 @@ import {
 } from './actions';
 
 jest.mock(
-  'shared/services/auth/services/state-token-generator/state-token-generator'
+  'shared/services/auth/services/random-string-generator/random-string-generator'
 );
 jest.mock('shared/services/api/api');
 jest.mock('shared/services/map-categories');
@@ -49,7 +49,7 @@ describe('containers/App/saga', () => {
   let origSessionStorage;
 
   beforeEach(() => {
-    stateTokenGenerator.mockImplementation(
+    randomStringGenerator.mockImplementation(
       () => 'n8vd9fv528934n797cv342bj3h56'
     );
     global.window.open = jest.fn();
@@ -103,7 +103,7 @@ describe('containers/App/saga', () => {
     });
 
     it('should dispatch error', () => {
-      stateTokenGenerator.mockImplementationOnce(() => undefined);
+      randomStringGenerator.mockImplementationOnce(() => undefined);
 
       const action = { payload };
